refactor: update gorm version

+++ controller/advertisement.go
@@ -180,7 +180,7 @@ func (ctrl advertisementCtrl) list(c *elton.Context) (err error) {
 	if err != nil {
 		return
 	}
-	count := -1
+	count := int64(-1)
 	args := params.toConditions()
 	queryParams := params.toPGQueryParams()
 	if queryParams.Offset == 0 {
@@ -196,7 +196,7 @@ func (ctrl advertisementCtrl) list(c *elton.Context) (err error) {
 	c.CacheMaxAge("1m")
 	c.Body = &struct {
 		Advertisements []*service.Advertisement `json:"advertisements,omitempty"`
-		Count          int                      `json:"count,omitempty"`
+		Count          int64                    `json:"count,omitempty"`
 	}{
 		result,
 		count,
